requirements.txt
appdirs==1.4.4
arrow==0.16.0
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.0.8
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.11.1
filelock==3.0.12
pdfminer.six==20200517
pipenv==2020.6.2
pycryptodome==3.9.8
Pygments==2.7.0
PySimpleGUI==4.24.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
virtualenv==20.0.26
virtualenv-clone==0.5.4
